Use simple vectors of (unsigned-byte 8) for wire
wire send and recv needs to use byte vectors.
wire-serialize and wire-deserialize are responsible for
converting to strings.

This fixes the UTF8 conversion error that was occurring when
a ZMQ provided identity with the structure #(0 w x y z)
where w x y z are bytes of a random 4-byte int.  identities should
never be encoded so the encoding needed to be moved out of the
message-send and message-recv functions into the wire-(de)serialize functions